Rename IGRF_C to mag_lookup
