+++ determination/mag_lookup/mag_lookup.h
